@ Xeon: Phase 7 T7.2+T7.3 complete - batch GEMM with VNNI matmul + result scatter
- T7.2: Per-expert batch GEMM: gather norms → quantize int8 → gate/up matmul
  (VPDPBUSD via ds4_xeon_matmul_a8w8_vnni_batch) → SiLU → quantize int16 →
  down matmul (VPDPWSSD via ds4_xeon_matmul_a16w16_vnni_batch)
- T7.3: Scatter weighted expert outputs back to per-token MoE accumulator
- T3.3.1 bugfix: 32-bit overflow in buffer size calc (n_expert*2*n_embd*n_ff_exp)
  caused 0-byte allocation → fixed with size_t casts
- Added batch matmul declarations to ds4_xeon.h
- Q2_K down dequant still scalar (~80% of per-layer dequant time);
  vectorizing Q2_K would reduce per-layer dequant from ~12s to ~2s

+++ ds4_xeon.c
@@ -1167,11 +1167,11 @@ int ds4_xeon_predequant_init(
     out->n_embd = n_embd;
     out->n_ff_exp = n_ff_exp;
 
-    // Compute per-buffer sizes
-    // gate_up: n_expert × 2(gate+up) × n_embd × n_ff_exp uint8
-    out->gate_up_bytes = (size_t)n_expert * 2 * n_embd * n_ff_exp;
-    // down: n_expert × n_ff_exp × n_embd int16
-    out->down_bytes = (size_t)n_expert * n_ff_exp * n_embd * sizeof(int16_t);
+    // Compute per-buffer sizes (cast to size_t to avoid 32-bit overflow)
+    // gate_up: n_expert × 2(gate+up) × n_embd × n_ff_exp uint8 = 4.0 GiB
+    out->gate_up_bytes = (size_t)n_expert * 2 * (size_t)n_embd * (size_t)n_ff_exp;
+    // down: n_expert × n_ff_exp × n_embd int16 = 4.0 GiB
+    out->down_bytes = (size_t)n_expert * (size_t)n_ff_exp * (size_t)n_embd * sizeof(int16_t);
 
     // Single buffer (per-layer dequant, 8.6 GB)
     for (int b = 0; b < 1; b++) {
